% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_image_square_to_circle.R
\name{convert_image_square_to_circle}
\alias{convert_image_square_to_circle}
\title{Title}
\usage{
convert_image_square_to_circle(image_path)
}
\arguments{
\item{image_url}{}
}
\description{
Title
}
