% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_df.R
\name{generate_df}
\alias{generate_df}
\title{Generate pseudorandom dataframes with specific parameters}
\usage{
generate_df(n = 10L, n_grps = 1L, with_seed = NULL, mean = c(10), sd = mean/10)
}
\arguments{
\item{n}{An integer indicating the number of rows per group, default to \code{10}}

\item{n_grps}{An integer indicating the number of rows per group, defaults to \code{1}}

\item{with_seed}{A seed to make the randomization reproducible}

\item{mean}{A number indicating the mean of the randomly generated values, must be a vector of equal length to the \code{n_grps}}

\item{sd}{A number indicating the standard deviation of the randomly generated values, must be a vector of equal length to the \code{n_grps}}
}
\value{
a tibble/dataframe
}
\description{
This function is a small utility to create a specific length dataframe
with a set number of groups, specific mean/sd per group. Note that the total length
of the dataframe will be \code{n} * \code{n_grps}.
}
\examples{
generate_df()
generate_df(n = 100L, n_grps = 5L, with_seed = NULL, mean = seq(10, 50, length.out = 5))

library(dplyr)
generate_df(
  100L,
  n_grps = 5,
  mean = seq(10, 50, length.out = 5)
) |>
  group_by(grp) |>
  summarise(
    mean = mean(values), # mean is approx mean
    sd = sd(values),     # sd is approx sd
    n = n(),             # each grp is of length n
    # showing that the sd default of mean/10 works
    `mean/sd` = round(mean / sd, 1)
  )
}
\seealso{
Other Utilities: 
\code{\link{gt_theme_ms}()}
}
\concept{Utilities}
