% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palattecode.R
\name{ms_brand_palette}
\alias{ms_brand_palette}
\title{Use Custom Brand Color Palette.}
\usage{
ms_brand_palette(palette = "main", ...)
}
\description{
Use Custom Brand Color Palette.
}
\examples{
<<<<<<< HEAD
<<<<<<< HEAD
library(automaton)
library(scales)
=======
>>>>>>> 3771dc7000c54237e34c99c812a36266aa80d7e6
=======
>>>>>>> 3771dc7000c54237e34c99c812a36266aa80d7e6
#The `show_col` function from the `scales` package is a nifty way to showcase all the colors available in a given palette:
scales::show_col(ms_brand_palette("main"), cex_label = 1)
scales::show_col(ms_brand_palette("highlight"), cex_label = 1)
}
